archive/backend_legacy/requirements-updated.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pymongo==4.8.0
motor==3.5.1
pydantic==2.8.2
redis>=4.5.0
celery>=5.3.0
python-bson>=0.23.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
bleach>=6.1.0
html2text>=2020.1.16

# Additional dependencies found in code analysis
openai>=1.0.0
bcrypt>=4.0.0
jwt>=1.3.1
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.1.0
qrcode>=7.4.0
reportlab>=4.0.0
Pillow>=9.0.0
aiohttp>=3.8.0
aiosmtplib>=2.0.0
cryptography>=3.4.0
jinja2>=3.1.0
python-multipart>=0.0.9
email-validator>=1.3.0
starlette>=0.27.0
websockets>=11.0.0
tornado>=6.2.0
gunicorn>=20.1.0
rich>=13.0.0
click>=8.1.0
yaml>=6.0.0
xlsxwriter>=3.1.0

# Development and testing
coverage>=7.2.0
pytest-cov>=4.0.0
factory-boy>=3.2.0
faker>=18.0.0

# Database and storage
gridfs>=0.9.0
pymongo[srv]>=4.8.0

# Security and validation
pydantic-settings>=2.0.0
python-dateutil>=2.8.0
tzdata>=2023.3
